backend/requirements.txt
# Django - Updated 2025-12-27
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Authentication
djangorestframework-simplejwt>=5.3
PyJWT>=2.8

# Database (PostgreSQL for production)
psycopg2-binary>=2.9
dj-database-url>=2.1

# Production server
gunicorn>=21.0
whitenoise>=6.6
gevent>=24.0

# Redis for caching
redis>=5.0
django-redis>=5.4

# Database connection pooling
django-db-connection-pool>=1.2

# Telegram notifications
python-telegram-bot>=21.0

# Utilities
python-dotenv>=1.0
pillow>=10.0
requests>=2.31

# Screenshot capture for daily reports
playwright>=1.40

# Cloud Storage (Cloudinary)
cloudinary>=1.36
django-cloudinary-storage>=0.3

# Testing
pytest-mock>=3.12.0
